src/departure_api/requirements.txt
numpy==1.15.1
pandas==0.23.4
pyproj==1.9.5.1
python-dateutil==2.7.3
requests==2.19.1
s3fs==0.1.6
scipy==1.1.0
Shapely==1.6.4.post2
smart-open==1.7.1
flask-restplus==0.11.0
gunicorn==19.9.0
